safety browserControl
